pica: Add pica_types module and move float24 definition.
