+++ src/video_core/CMakeLists.txt
@@ -33,6 +33,7 @@ set(HEADERS
             command_processor.h
             gpu_debugger.h
             pica.h
+            pica_types.h
             primitive_assembly.h
             rasterizer.h
             rasterizer_interface.h
